requirements.txt
Flask
supabase-py
bcrypt
python-dotenv
qrcode[pil]
Pillow

